+++ tests/testthat/test-combine.R
@@ -11,6 +11,27 @@ test_that('combine and funnel store history in list', {
   expect_equal(length(as.list(combine(l))[-1]), 2)
   expect_equal(length(as.list(funnel("hi", 42))[-1]), 2)
 
+  # basic anonymous functions work
+  expect_equal({
+    "a" %>>%
+       funnel(b="b", c="c") %*>%
+       (function(a,b,c) {paste(a,b,c)}) %>% esc 
+  }, "a b c" )
+
+  # keyword arguments are respected
+  expect_equal({
+    "a" %>>%
+       funnel(c="c", b="b") %*>%
+       (function(a,b,c) {paste(a,b,c)}) %>% esc 
+  }, "a b c" )
+
+  # The value from the pipe defaults to first position
+  expect_equal({
+    "a" %>>%
+       funnel("c", "b") %*>%
+       (function(a,b,c) {paste(a,b,c)}) %>% esc 
+  }, "a c b" )
+
 })
 
 test_that('combine and funnel store all results into an order preserving list', {